backend/requirements.txt
Flask>=2.2
bcrypt>=4.0
cryptography>=41.0
python-dotenv>=1.0
requests>=2.31
itsdangerous>=2.1
google
google-genai